Extracted EvileItemTags. Updated references from EnumEvileArmorMaterial
